Tweak the msan config
The memory sanitizer config option now successfully compiles the project but
unfortunately we cannot run our tests with it due to some gtest issues.
Apparently it requires we use libc++ and compile it under msan.

The tsan (thread sanitizer) and ubsan (undefined behavior sanitizer)
configs both work and are able to run our tests successfully and the
main program. A nice improvement would be to add them as steps to the
travis build