Remove registry specific Id creation
Previously the way we were creating `Id`s was to use
`registry->CreateId(...)` with the idea of being able to have `Id`s that
were customized to a particular registry. This hasn't been needed at all
and it just makes things too verbose. Switch the creation to
`Id::Of(name, [tags])` instead.