New payload fmt and add monotonically sampled ctrs
Update the payload format by removing the protocol version and
introducing tags with a comma instead of `:#`

In order to help implement some sidecars that are relaying samples from
other sources, we add a new meter type which is a monotonic source that
was sampled at a given time in the recent past using the type `X`. This
is experimental and subject to be removed based on feedback.

`X,1000:some.name:1`
`X,2000:some.name:2`