Fix bug when copying large tag values (#27)
We are using a small buffer optimization for tags where we do not need
to allocate memory if the number of tags is <= 8. We are saving memory
by reusing the array of tags to store the capacity when the tags live on
the heap. However we were not initializing the capacity correctly when
copying/moving from another large Tags value.