add --enable_socket flag to allow for disabling the unix socket
We have some teams that would prefer to run this service without the UNIX
domain socket. To preserve backward compatibility, the default value of this
flag is true.

The handling of enable flags has been adjusted slightly, so that optionals
are used, which makes the code a little easier to read when passing these
values around.

A few functions which are no longer needed since the switch to the Abseil
flags library have been removed.

A Dockerfile and build script have been added to the project, to provide a
reproducible means of running builds locally. To use this build script, you
must provide a BASEOS_IMAGE environment variable, which should resolve to a
reasonable image, such as an Ubuntu LTS version.

In order to allow the build to complete successfully, it is recommended to
configure Docker Desktop with 8GB RAM.