add --age_gauge_limit flag
This change implements a limit on the maximum number of age gauges that may
be reported from a spectatord process.

The purpose of this feature is to limit a potential source of memory leaks
for long-running spectatord processes where the tags associated with an age
gauge vary over time. The limit defaults to 1000 and we do not think that
most users will hit this limit on a per-node basis. The limit is customizable
with the --age_gauge_limit flag, in the event that users need to adjust this
setting for their particular service.

The first time that the age gauge limit is exceeded, the full list of known
age gauges will be logged, to assist with operator debugging.