fix local server startup
This bug was introduced with the PR that added the --enable_unix_socket
flag. The creation of the server was moved into the scope of an if
statement, which meant that it was not available to the io context
when it started, which produced the following errors:

[info] Starting local server (dgram) on socket /run/spectatord/spectatord.unix
[error] Error receiving: 125: Operation aborted.
[error] Error receiving: 9: Bad file descriptor

This change follows the model of the statsd server, by declaring a unique
pointer to the local server outside the scope of the if statement.