improve build script ergonomics
Since the script has been reconfgured to use a local .cache directory, which
improves rebuild times, the symlinks added at the end of the bazel runs were
incorrect for local use.

This change adjusts the symlink locations so that they make sense outside of
the container, while still being relevant inside the container.

The `shell` option for the build script now has a workdir option set, which
starts the session in the source directory.