do not abort when disabled
If an instance had a disabled file defined, then spectatord would never
send metrics, which would cause it to abort every minute and fill up
available disk space with core files.

This change considers whether or not spectatord is enabled when checking the
last time since it was able to send metrics successfully.

Added a trace log for the condition where we skip sending metrics, to make
it more obvious what is happening when verbose logging is enabled.