remove libbfd configuration
The addition of libbfd is intended to provide prettier stack traces with
backwards-cpp. Linking against libbfd has never worked correctly for this
project and it complicates the build, so remove it and stick with the
basic stack tracing enabled by the header-only implementation of backwards.

Full static linking of the main binary results in segfaults, due to the
use of libnss, so drop back to only static linking libstdc++, which is
necessary for running on Titus.