admin server
This change adds an administrative server to SpectatorD, for the purpose
of providing debugging information and making small changes to the running
configuration.

The Mantis application needed the ability to change their common tags on
the fly during operation, as they swap out jobs and the `/config/common_tags`
endpoint provides this feature.

Users of AgeGauges occasionally need the ability to delete the AgeGauges they
create - the `/metrics/A` endpoint provides this feature.

The version of Conan is locked to the latest release before 2.0, to avoid
breaking changes for the time being.