revert libbfd linking for backward-cpp (#69)
Linking against this library introduces strong distro-specific requirements
which complicates the build and distribution process. We want to keep things
simple and as generic as possible, so we will abandon getting ultimate stack
traces on the Linux platform.

To help deal with segmentation faults, we will press forward with migrating
to systemd startup scripts, so that we can control the behavior of core dumps.