reduce spectatord logging
The `Sent:` messages serve as a positive confirmation that spectatord is
either delivering or dropping measurements, but there is generally no action
to be taken from it directly, and this information is also recorded in the
`spectator.measurements` metric.

In the case of Titus, this message alone is responsible for 25% of all logs.

Moving this message to the `debug` level drops it from the standard use case,
while still allowing it to be accessed if the `--verbose` flag is passed to
the binary.

Fixes #71.