add --no_common_tags flag
The purpose of this flag is to remove all common tags, so that users can
override any common tags they wish, including `nf.app`, by providing them
in the protocol line. Note that a proper environment is still required, so
that the publish URI can be determined correctly.

This feature should only be used for secondary spectatord processes on an
instance. A standard process should exist, so that atlas-system-agent metrics
can be received and published.

When this feature is active, the publishing of internal status metrics is
disabled, because we cannot attribute the metrics correctly in the absence
of common tags.