add --ipv4_only parameter (default = false)
There are some systems which cannot run IPv6 at all, and the previous default
behavior of spectatord meant that the process crashes when attempting to listen
on a udp6 socket.

This change adds an optional parameter --ipv4_only, defaulting to false, which
will configure the socket listener to use udp4 instead of udp6.

When using the udp6 socket, the server is capable of receiving udp packets from
IPv4 on dual-stacked instances with IPv4-mapped IPv6 addresses enabled.