log parse errors (#102)
There was no feedback mechanism for poorly formatted protocol messages, because
the errors were not logged. This change logs those errors, at INFO level, so
that we do not trigger exception alerts.

External publishing now needs a slightly longer connect timeout, in order for
connections to succeed. Update the netflix-spectatord-cppconf library to make
this happen.