reduce logging levels for http errors
We see occasional HTTP errors when POSTing data to the backend. These are
often transient errors, and there is nothing that an end user can do about
them. Since we are ingesting spectatord logs, this results in creating
exception alerts, and we do not want this to happen. Thus, we turn down the
log levels for HTTP errors to info level, to prevent this scenario.