fix `spectator.measurements` metrics
* The `spectator.measurements` internal status metrics were not being published
on 200 OK responses, only when partial errors were observed. This change fixes
that, so that we always publish the sent metrics counter on 200 OK.
* The `spectator.measurements` metrics were using the `owner` tag to identify
the source of the metrics. Our internal dashboards use the `nf.process` tag for
this purpose, so we want to use that instead.
* The `ipc.client.call` metrics were using the `owner` tag to identify the source
of the metrics. The `owner` tag is better considered to be a way to identify
a source library, when there could be many sources of a metric. It is better to
use `nf.process` to identify these metrics.
* The `spectator.registrySize` metrics were using the `owner` tag to identify the
source of the metrics. This was swapped to `nf.process`.
* We do not want to set the `nf.process` tag with an environment variable for
`spectatord`, because there are many processes that send metrics through it,
each of which may want to set that tag independently. Thus, we set the
`nf.process` tag manually in the status metrics.
* The internal configuration library version was updated, so that we default to
the test endpoint for external publishing.
* The command line parameter parsing was updated, so that the `--uri` parameter
can be used to switch back to the prod endpoint, for external publishing, if
there is a need.